chore: rename walletClient.ts -> client.ts
