+++ packages/notify-client/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./constants";
 export * from "./types";
 
-export { NotifyClient } from "./walletClient";
+export { NotifyClient } from "./client";
